Move to unsigned integers, fixed warnings

+++ src/vectors/vector.c
@@ -15,12 +15,12 @@
 
 #define INLINE 
 
-INLINE int
-internal_ceil_log_2(int size)
+INLINE unsigned int
+internal_ceil_log_2(unsigned int size)
 {
     size--;
     if (size <= 0) return 0;
-    for (int result = 0; ; size >>= 4) {
+    for (unsigned int result = 0; ; size >>= 4) {
         if ((size >> 0) == 0) return result;
         if ((size >> 1) == 0) return result + 1;
         if ((size >> 2) == 0) return result + 2;
@@ -32,21 +32,21 @@ internal_ceil_log_2(int size)
     return 0;
 }
 
-INLINE int
-pow_of_2(int power)
+INLINE unsigned int
+pow_of_2(unsigned int power)
 {
-    return 1 << power;
+    return 1U << power;
 }
 
 inline struct vector*
-create_vector(const int capacity, const int member_size)
+create_vector(const unsigned int capacity, const unsigned int member_size)
 {
-    int alloc_capacity = capacity;
+    unsigned int alloc_capacity = capacity;
     if (alloc_capacity <= 0) alloc_capacity = DEFAULT_VECTOR_CAPACITY;
     
     struct vector* result = malloc(sizeof(struct vector));
-    int log2_member_size = internal_ceil_log_2(member_size);
-    void* data = malloc((size_t)alloc_capacity << (size_t)log2_member_size);
+    unsigned int log2_member_size = internal_ceil_log_2(member_size);
+    void* data = malloc(alloc_capacity << log2_member_size);
 
     *result = (struct vector) {
         .capacity = alloc_capacity,
@@ -62,13 +62,13 @@ create_vector(const int capacity, const int member_size)
 }
 
 inline struct vector*
-create_vector_cleared(const int capacity, const int member_size)
+create_vector_cleared(const unsigned int capacity, const unsigned int member_size)
 {
     ASSERT(capacity > 0 && "Need positive, non-zero capacity");
     
     struct vector* result = malloc(sizeof(struct vector));
-    int log2_member_size = internal_ceil_log_2(member_size);
-    void* data = calloc((size_t)capacity << (size_t)log2_member_size, sizeof(char));
+    unsigned int log2_member_size = internal_ceil_log_2(member_size);
+    void* data = calloc(capacity << log2_member_size, sizeof(char));
 
     *result = (struct vector) {
         .capacity = capacity,
@@ -88,9 +88,9 @@ vector_fill_value(struct vector* vv, void* fill)
 {
     vv->size = vv->capacity;
     if (fill != NULL) {
-        for (int ii = 0; ii < vv->capacity; ii++) {
-            memcpy((char*)vv->data + (size_t)(ii << vv->member_size_log2), fill,
-                    (size_t)vv->member_size);
+        for (unsigned int ii = 0; ii < vv->capacity; ii++) {
+            memcpy((char*)vv->data + (ii << vv->member_size_log2), fill,
+                    vv->member_size);
         }
     }
     return vv;
@@ -106,25 +106,25 @@ vector_fill(struct vector* vv)
 inline int
 destroy_vector(struct vector* vv)
 {
-    memset(vv->data, 0, (size_t)vv->capacity << vv->member_size_log2);
+    memset(vv->data, 0, vv->capacity << vv->member_size_log2);
     free(vv->data);
     memset(vv, 0, sizeof(*vv));
     free(vv);
     return 0;
 }
 
-inline int
+inline unsigned int
 vector_size(struct vector* vv)
 {
     return vv->size;
 }
 
 inline struct vector*
-vector_resize(struct vector* vv, const int new_capacity)
+vector_resize(struct vector* vv, const unsigned int new_capacity)
 {
     ASSERT(new_capacity > 0);
 
-    vv->data = realloc(vv->data, ((size_t)new_capacity) << (size_t)vv->member_size_log2);
+    vv->data = realloc(vv->data, (new_capacity) << vv->member_size_log2);
 
     ASSERT(vv->data != NULL);
 
@@ -134,7 +134,7 @@ vector_resize(struct vector* vv, const int new_capacity)
 }
 
 inline int
-vector_get(struct vector* vv, const int index, void* ret)
+vector_get(struct vector* vv, const unsigned int index, void* ret)
 {
     void* ref = NULL;
 
@@ -142,16 +142,16 @@ vector_get(struct vector* vv, const int index, void* ret)
     (void)result;
     ASSERT(result == 0);
     if (ref == NULL) {
-        memset(ret, 0, (size_t)pow_of_2(vv->member_size_log2));
+        memset(ret, 0, pow_of_2(vv->member_size_log2));
         return 1;
     }
-    memcpy(ret, ref, (size_t)vv->member_size);
+    memcpy(ret, ref, vv->member_size);
 
     return 0;
 }
 
 inline int
-vector_set(struct vector* vv, const int index, const void* val)
+vector_set(struct vector* vv, const unsigned int index, const void* val)
 {
     void* ref = NULL;
 
@@ -161,13 +161,13 @@ vector_set(struct vector* vv, const int index, const void* val)
     if (ref == NULL) {
         return 1;
     }
-    memcpy(ref, val, (size_t)vv->member_size);
+    memcpy(ref, val, vv->member_size);
 
     return 0;
 }
 
 inline int
-vector_get_ref(struct vector* vv, const int index, void** ret)
+vector_get_ref(struct vector* vv, const unsigned int index, void** ret)
 {
     if (index >= vv->size) return 1;
     
@@ -180,13 +180,13 @@ inline struct vector*
 vector_push(struct vector* vv, void* data)
 {
     if (vv->size == vv->capacity) {
-        vv->data = realloc((char*)vv->data, 2 * (size_t)vv->capacity << (size_t)vv->member_size_log2);
+        vv->data = realloc((char*)vv->data, 2 * vv->capacity << vv->member_size_log2);
         ASSERT(vv != NULL);
         vv->capacity *= 2;
     }
     
     void* ref = (char*)vv->data + (vv->size << vv->member_size_log2);
-    memcpy(ref, data, (size_t)vv->member_size);
+    memcpy(ref, data, vv->member_size);
     vv->size++;
 
     return vv;
@@ -197,8 +197,8 @@ vector_pop(struct vector* vv, void* ret)
 {
     vv->size--;
     if (ret != NULL) {
-        void* ref = (char*)vv->data + ((size_t)vv->size << vv->member_size_log2);
-        memcpy(ret, ref, (size_t)vv->member_size);
+        void* ref = (char*)vv->data + (vv->size << vv->member_size_log2);
+        memcpy(ret, ref, vv->member_size);
     }
 
     return vv;
